@@ -8,16 +8,16 @@
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.8",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
     "@types/yup": "^0.29.13",
-    "bootstrap": "^5.1.1",
+    "bootstrap": "^5.2.2",
     "formik": "^2.2.9",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.5.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.8.4",
     "web-vitals": "^1.1.2",
     "yup": "^0.32.9"
   },
